requirements.txt
google-adk>=0.1.0
google-generativeai>=0.3.0
google-cloud-bigquery>=3.10.0
chromadb>=0.4.0
duckduckgo-search>=4.0.0
db-dtypes
pandas
python-dotenv
